+++ docs/src/markdown/about/changelog.md
@@ -4,6 +4,7 @@
 
 - **FIX**: CSS selectors should be evaluated with CSS whitespace rules.
 - **FIX**: Processing instructions, CDATA, and declarations should all be ignored from child considerations for `:empty`.
+- **FIX**: In Beautiful Soup, the document itself is the first tag. Do not match the "document" tag by returning false for any tag that doesn't have a parent.
 
 ## 1.0.0b1